Create Assignment_3_code
The code was implemented and executed in steps.
A basic approach was initially used utilising RTOS documentation from the online resource. RTOS was implemented by initialising null tasks and task handles.
An Additional approach was utilised to suspend, pause and delete tasks, but this proved less effective than utilising a Queue method. The remainder of tasks did not need a queue however and was implemented by utilising the inbuilt task scheduler, task delays and semaphores to distribute data between the three tasks.